test: refactor tests to provide unique test per source/destination
