+++ rudderstack/integrations/destinations/destination_postgres.go
@@ -1,4 +1,4 @@
-package configs
+package destinations
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -81,15 +81,10 @@ func init() {
 				Required:         true,
 				ValidateDiagFunc: c.StringMatchesRegexp("^(disable|require)$"),
 			},
-
 			"use_rudder_storage": {
 				Type:     schema.TypeBool,
 				Optional: true,
 				Default:  false,
-			},
-			"bucket_provider": {
-				Type:     schema.TypeString,
-				Optional: true,
 			},
 			"sync": {
 				Type:     schema.TypeList,
@@ -255,66 +250,6 @@ func init() {
 						},
 					},
 				},
-			},
-		},
-		TestConfigs: []c.TestConfig{
-			{
-				TerraformCreate: `
-					host = "example.com"
-					port = "5432"
-					user = "example-user"
-					password = "example-password"
-					namespace = "example-namespace"
-					ssl_mode = "disable"
-					database = "example-database"
-					use_rudder_storage = true
-					s3 {
-						bucket_name = "some-bucket-name"
-						access_key_id = "some-access-key-id"
-						access_key = "some-access-key"
-					}
-					`,
-				APICreate: `{
-					"host": "example.com",
-					"port": "5432",
-					"user": "example-user",
-					"password": "example-password",
-					"database": "example-database",
-					"namespace": "example-namespace",
-					"sslMode": "disable",
-					"useRudderStorage": true,
-					"bucketProvider": "S3",
-					"bucketName": "some-bucket-name",
-					"accessKeyID": "some-access-key-id",
-					"accessKey": "some-access-key"
-				}`,
-				TerraformUpdate: `
-					host = "example.com"
-					port = "5432"
-					user = "example-user-updated"
-					password = "example-password"
-					namespace = "example-namespace"
-					ssl_mode = "disable"
-					database = "example-database"
-					use_rudder_storage = true
-					gcs {
-						bucket_name = "some-bucket-name"
-						credentials = "some-credentials"
-					}
-					`,
-				APIUpdate: `{
-					"host": "example.com",
-					"port": "5432",
-					"user": "example-user-updated",
-					"password": "example-password",
-					"database": "example-database",
-					"namespace": "example-namespace",
-					"sslMode": "disable",
-					"useRudderStorage": true,
-					"bucketProvider": "GCS",
-					"bucketName": "some-bucket-name",
-					"credentials": "some-credentials"
-				}`,
 			},
 		},
 	})
