+++ rudderstack/integrations/sources/sources.go
@@ -1,18 +1,16 @@
-package configs
+package sources
 
 import c "github.com/rudderlabs/terraform-provider-rudderstack/rudderstack/configs"
 
 func init() {
 	c.Sources.Register("http", c.ConfigMeta{
-		APIType:     "HTTP",
-		Properties:  []c.ConfigProperty{},
-		SkipConfig:  true,
-		TestConfigs: []c.TestConfig{c.EmptyTestConfig},
+		APIType:    "HTTP",
+		Properties: []c.ConfigProperty{},
+		SkipConfig: true,
 	})
 	c.Sources.Register("javascript", c.ConfigMeta{
-		APIType:     "Javascript",
-		Properties:  []c.ConfigProperty{},
-		SkipConfig:  true,
-		TestConfigs: []c.TestConfig{c.EmptyTestConfig},
+		APIType:    "Javascript",
+		Properties: []c.ConfigProperty{},
+		SkipConfig: true,
 	})
 }
